+++ engine/runtime/src/test/scala/org/enso/interpreter/test/instrument/RuntimeSuggestionUpdatesTest.scala
@@ -36,7 +36,7 @@ class RuntimeSuggestionUpdatesTest
     val tmpDir: File = Files.createTempDirectory("enso-test-packages").toFile
 
     val pkg: Package[File] =
-      PackageManager.Default.create(tmpDir, packageName, "0.0.1")
+      PackageManager.Default.create(tmpDir, packageName, "Enso_Test")
     val out: ByteArrayOutputStream = new ByteArrayOutputStream()
     val executionContext = new PolyglotContext(
       Context
@@ -116,10 +116,10 @@ class RuntimeSuggestionUpdatesTest
   it should "send suggestion updates after file modification" in {
     val contextId  = UUID.randomUUID()
     val requestId  = UUID.randomUUID()
-    val moduleName = "Test.Main"
+    val moduleName = "Enso_Test.Test.Main"
 
     val code =
-      """from Builtins import all
+      """from Standard.Builtins import all
         |
         |main = IO.println "Hello World!"
         |""".stripMargin.linesIterator.mkString("\n")
@@ -145,7 +145,7 @@ class RuntimeSuggestionUpdatesTest
         Api.PushContextRequest(
           contextId,
           Api.StackItem.ExplicitCall(
-            Api.MethodPointer(moduleName, "Test.Main", "main"),
+            Api.MethodPointer(moduleName, "Enso_Test.Test.Main", "main"),
             None,
             Vector()
           )
@@ -183,9 +183,9 @@ class RuntimeSuggestionUpdatesTest
                     "main",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None)
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -220,7 +220,7 @@ class RuntimeSuggestionUpdatesTest
         Api.SuggestionsDatabaseModuleUpdateNotification(
           file = mainFile,
           version = contentsVersion(
-            """from Builtins import all
+            """from Standard.Builtins import all
               |
               |main =
               |    x = 42
@@ -238,9 +238,9 @@ class RuntimeSuggestionUpdatesTest
                     "main",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None)
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -296,7 +296,7 @@ class RuntimeSuggestionUpdatesTest
         Api.SuggestionsDatabaseModuleUpdateNotification(
           file = mainFile,
           version = contentsVersion(
-            """from Builtins import all
+            """from Standard.Builtins import all
               |
               |main =
               |    x = 42
@@ -315,9 +315,9 @@ class RuntimeSuggestionUpdatesTest
                     "main",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None)
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -392,7 +392,7 @@ class RuntimeSuggestionUpdatesTest
         Api.SuggestionsDatabaseModuleUpdateNotification(
           file = mainFile,
           version = contentsVersion(
-            """from Builtins import all
+            """from Standard.Builtins import all
               |
               |main =
               |    x = 42
@@ -412,9 +412,9 @@ class RuntimeSuggestionUpdatesTest
                     "main",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None)
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -497,7 +497,7 @@ class RuntimeSuggestionUpdatesTest
         Api.SuggestionsDatabaseModuleUpdateNotification(
           file = mainFile,
           version = contentsVersion(
-            """from Builtins import all
+            """from Standard.Builtins import all
               |
               |foo x = x * 10
               |
@@ -519,9 +519,9 @@ class RuntimeSuggestionUpdatesTest
                     "main",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None)
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -584,11 +584,11 @@ class RuntimeSuggestionUpdatesTest
                     "foo",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None),
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None),
                       Suggestion
                         .Argument("x", Constants.ANY, false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -623,7 +623,7 @@ class RuntimeSuggestionUpdatesTest
         Api.SuggestionsDatabaseModuleUpdateNotification(
           file = mainFile,
           version = contentsVersion(
-            """from Builtins import all
+            """from Standard.Builtins import all
               |
               |foo a b = a * b
               |
@@ -645,11 +645,11 @@ class RuntimeSuggestionUpdatesTest
                     "foo",
                     List(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None),
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None),
                       Suggestion
                         .Argument("x", Constants.ANY, false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),
@@ -685,10 +685,10 @@ class RuntimeSuggestionUpdatesTest
   it should "index overloaded functions" in {
     val contextId  = UUID.randomUUID()
     val requestId  = UUID.randomUUID()
-    val moduleName = "Test.Main"
+    val moduleName = "Enso_Test.Test.Main"
 
     val contents =
-      """from Builtins import all
+      """from Standard.Builtins import all
         |
         |main =
         |    x = 15.overloaded 1
@@ -721,7 +721,7 @@ class RuntimeSuggestionUpdatesTest
         Api.PushContextRequest(
           contextId,
           Api.StackItem.ExplicitCall(
-            Api.MethodPointer(moduleName, "Test.Main", "main"),
+            Api.MethodPointer(moduleName, "Enso_Test.Test.Main", "main"),
             None,
             Vector()
           )
@@ -759,9 +759,9 @@ class RuntimeSuggestionUpdatesTest
                     "main",
                     Seq(
                       Suggestion
-                        .Argument("this", "Test.Main", false, false, None)
+                        .Argument("this", "Enso_Test.Test.Main", false, false, None)
                     ),
-                    "Test.Main",
+                    "Enso_Test.Test.Main",
                     Constants.ANY,
                     None
                   ),