+++ engine/runtime/src/test/scala/org/enso/interpreter/test/semantic/RuntimeManagementTest.scala
@@ -21,7 +21,7 @@ class RuntimeManagementTest extends InterpreterTest {
         .asHostObject[Context]()
 
       val code =
-        """from Builtins import all
+        """from Standard.Builtins import all
           |
           |foo x =
           |    if x == 0 then IO.println "Start." else Nothing
@@ -75,7 +75,7 @@ class RuntimeManagementTest extends InterpreterTest {
     "Automatically free managed resources" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |type Mock_File i
           |
@@ -113,7 +113,7 @@ class RuntimeManagementTest extends InterpreterTest {
     "Automatically free managed resources amongst manual closure of other managed resources" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |type Mock_File i
           |
@@ -152,7 +152,7 @@ class RuntimeManagementTest extends InterpreterTest {
     "Automatically free managed resources amongst manual takeover of other managed resources" in {
       val code =
         """
-          |from Builtins import all
+          |from Standard.Builtins import all
           |
           |type Mock_File i
           |